requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
pymongo[srv]==4.7.3
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
